% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVideo.R
\name{createVideo}
\alias{createVideo}
\title{Create video}
\usage{
createVideo(server, tests, comparisonEndPoint = c("visual", "all", "doc",
  "full"), dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{tests}{A character vector of test ids}

\item{comparisonEndPoint}{frame comparison end point: 
[visual]=visually complete | all=last change 
| doc=document complete | full=fully loaded}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A list of relevant video url's
}
\description{
create a video from <tests> (character vector of test ids)
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, days = 10L, filter = "google.com")
  tests <- tail(history[["Test ID"]], 5)
  createVideo(WPT, tests = tests)
}
}

