% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmbedVideoPlayer.R
\name{getEmbedVideoPlayer}
\alias{getEmbedVideoPlayer}
\title{Get embedded video player}
\usage{
getEmbedVideoPlayer(server, testId, file = NULL, dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{file}{Write the result returned to file. Defaults to NULL
= don't write to file}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
Writes a html file with embedded video if file argument is not
NULL else returns HTML.
}
\description{
get a html5 player for a video <id>
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, days = 10L, filter = "google.com")
  tests <- tail(history[["Test ID"]], 1)
  vidRes <- getEmbedVideoPlayer(WPT, tests$testId)
}
}

