% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoogleCsiData.R
\name{getGoogleCsiData}
\alias{getGoogleCsiData}
\title{get Google CSI data}
\usage{
getGoogleCsiData(server, testId, run = 1L, repeatView = FALSE,
  dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{run}{an integer. Which run number on a multiple runs test. 
Defaults to 1L.}

\item{repeatView}{get the Repeat View (cached view) instead of default 
First View (primed cache). Defaults to FALSE}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A tibble of Google CSI data (Client Side Instrumentation). If 
   no CSI data is available for the test a blank tibble is returned.
}
\description{
get Google CSI data (Client Side Instrumentation)
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, filter = "London")
  tID <- as.character(history[dim(history)[1], "Test ID"])
  googlecsi <- getGoogleCsiData(WPT, tID)
}
}

