% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHARData.R
\name{getHARData}
\alias{getHARData}
\title{Get HAR data}
\usage{
getHARData(server, testId, dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
returns a parsed HAR object from \code{\link[HARtools]{readHAR}}
}
\description{
get the HTTP Archive (HAR) from test
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, filter = "London")
  tID <- as.character(history[dim(history)[1], "Test ID"])
  har <- getHARData(WPT, tID)
}
}

