% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistory.R
\name{getHistory}
\alias{getHistory}
\title{Get test history}
\usage{
getHistory(server, days = 1L, dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{days}{Number of days to get history for defaults to 1L.}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A tibble with "Date/Time", "Location", "Test ID", "URL","Label" 
   of tests. NOTE: The results returned are limited on default server.
}
\description{
Get history of previously run tests
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, days = 7L)  
}
}

