% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetLogData.R
\name{getNetLogData}
\alias{getNetLogData}
\title{Get Net Log data (chrome)}
\usage{
getNetLogData(server, testId, run = 1L, repeatView = FALSE,
  dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{run}{an integer. Which run number on a multiple runs test. 
Defaults to 1L.}

\item{repeatView}{get the Repeat View (cached view) instead of default 
First View (primed cache). Defaults to FALSE}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A list containing the Net Log data
}
\description{
get the Chrome Developer Tools Net log data (if available) from test
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, filter = "London")
  tID <- as.character(history[dim(history)[1], "Test ID"])
  timelineData <- getNetLogData(WPT, tID)
}
}

