% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPageSpeedData.R
\name{getPageSpeedData}
\alias{getPageSpeedData}
\title{Get Page Speed results}
\usage{
getPageSpeedData(server, testId, run = 1L, repeatView = FALSE,
  dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{run}{an integer. Which run number on a multiple runs test. 
Defaults to 1L.}

\item{repeatView}{get the Repeat View (cached view) instead of default 
First View (primed cache). Defaults to FALSE}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A list of Google Page Speed results (if available) from test.
   It is unlikely pagespeed data is available for recent tests.
   See \url{https://github.com/WPO-Foundation/webpagetest/issues/669}.
   If the Page Speed data is not available a HTTP 404 not found error
   will be returned.
}
\description{
get the Google Page Speed results (if available) from test
}
\examples{
\dontrun{
  WPT <- server()
  tID <- "121216_XZ_AA"
  pagespeed <- getPageSpeedData(WPT, tID)
}
}

