% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequestData.R
\name{getRequestData}
\alias{getRequestData}
\title{Get request data}
\usage{
getRequestData(server, testId, run = 1L, repeatView = FALSE,
  dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{run}{an integer. Which run number on a multiple runs test. 
Defaults to 1L.}

\item{repeatView}{get the Repeat View (cached view) instead of default 
First View (primed cache). Defaults to FALSE}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A tibble containing the request data
}
\description{
get the request data from test
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, filter = "London")
  tID <- as.character(history[dim(history)[1], "Test ID"])
  requestData <- getRequestData(WPT, tID)
}
}

