% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreenshotImage.R
\name{getScreenshotImage}
\alias{getScreenshotImage}
\title{Get the Screenshot PNG image}
\usage{
getScreenshotImage(server, testId, run = 1L, file = NULL,
  repeatView = FALSE, thumbnail = FALSE, fullResolution = FALSE,
  startRender = FALSE, documentComplete = FALSE, dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{run}{an integer. Which run number on a multiple runs test. 
Defaults to 1L.}

\item{file}{Write the result returned to file. Defaults to NULL
= don't write to file}

\item{repeatView}{get the Repeat View (cached view) instead of default 
First View (primed cache). Defaults to FALSE}

\item{thumbnail}{logical. Get the thumbnail of actual image. Defaults to
FALSE}

\item{fullResolution}{logical. get full resolution screenshot in PNG 
format if available. Defaults to FALSE.}

\item{startRender}{logical. get the page screenshot at the Start Render point (i.e.: when something was first 
displayed on screen). Defaults to FALSE.}

\item{documentComplete}{logical. get the page screenshot at the 
Document Complete point (i.e.: when window.onload was fired) 
Defaults to FALSE.}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
If a file argument is given the JPG/PNG returned is written to 
   file. Else a raster is returned via
    \link[jpeg]{readJPEG}|\link[png]{readPNG}. If a file is not 
    available when calling an argument (for example fullres, 
    docComplete, ....) a HTTP 404 Not found error is returned.
}
\description{
get the fully loaded page screenshot in JPG format (PNG if in full 
   resolution)
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, filter = "London")
  tID <- as.character(history[dim(history)[1], "Test ID"])
  screenshot <- getScreenshotImage(WPT, tID)
}
}

