% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTesters.R
\name{getTesters}
\alias{getTesters}
\title{Get testers status}
\usage{
getTesters(server, requestId = NULL, dryrun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{requestId}{request ID, useful to track asynchronous requests}

\item{dryrun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A list of testers status and details
}
\description{
list testers status and details
}
\examples{
\dontrun{
  WPT <- server()
  testers <- getTesters(WPT)  
}
}

