% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestInfo.R
\name{getTestInfo}
\alias{getTestInfo}
\title{Get test info}
\usage{
getTestInfo(server, testId, dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
returns info on a test
}
\description{
get test request info/details
}
\examples{
\dontrun{
  WPT <- server()
  testinfo <- getTestInfo(WPT, testId = "someid")  
}
}

