% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUtilizationData.R
\name{getUtilizationData}
\alias{getUtilizationData}
\title{Get utilisation data}
\usage{
getUtilizationData(server, testId, run = 1L, repeatView = FALSE,
  dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{run}{an integer. Which run number on a multiple runs test. 
Defaults to 1L.}

\item{repeatView}{get the Repeat View (cached view) instead of default 
First View (primed cache). Defaults to FALSE}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A tibble of the CPU, bandwidth and memory utilisation data 
   from test
}
\description{
get the CPU, bandwidth and memory utilisation data from test
}
\examples{
\dontrun{
  WPT <- server()
  history <- getHistory(WPT, filter = "London")
  tID <- as.character(history[dim(history)[1], "Test ID"])
  utilizationData <- getUtilizationData(WPT, tID)
}
}

