% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{server}
\alias{server}
\title{WPT server}
\usage{
server(serverURL = "http://www.webpagetest.org")
}
\arguments{
\item{serverURL}{the WPT server URL defaults to 
<http://www.webpagetest.org>}
}
\value{
returns a WPTserver object to be used by other functions
}
\description{
Detail a WPT server for tests
}
\examples{
\dontrun{
myServer <- server()
myServer
}
}

