% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancelTest.R
\name{cancelTest}
\alias{cancelTest}
\title{Cancel test}
\usage{
cancelTest(server, testId, key = Sys.getenv("WPTKEY"), dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{key}{API key (if assigned). Contact the WebPageTest server 
administrator for a key if required}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
A
}
\description{
cancel running/pending test
}
\examples{
\dontrun{
  WPT <- server()
  locs <- getLocations(WPT)
  test <- runTest(WPT, "r-project.org", 
                  location = paste0(locs$`ec2-eu-west-1`, ":Chrome"))
  cancelTest(WPT, test$testId)
}
}

