% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocations.R
\name{getLocations}
\alias{getLocations}
\title{Get server locations}
\usage{
getLocations(server, requestId = NULL, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{requestId}{request ID, useful to track asynchronous requests}

\item{...}{Additonal function arguments}
}
\value{
A list of locations and the number of pending tests at those locations
}
\description{
List locations and the number of pending tests
}
\examples{
\dontrun{
  WPT <- server()
  locs <- getLocations(WPT)  
}
}

