% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestStatus.R
\name{getTestStatus}
\alias{getTestStatus}
\title{check test status}
\usage{
getTestStatus(server, testId, requestId = NULL, dryRun = FALSE, ...)
}
\arguments{
\item{server}{A WebPagetest server}

\item{testId}{test ID string}

\item{requestId}{request ID, useful to track asynchronous requests}

\item{dryRun}{If TRUE just return the RESTful API URL. Defaults to 
FALSE}

\item{...}{Additonal function arguments}
}
\value{
returns the status of a test
}
\description{
check the status of a test
}
\examples{
\dontrun{
  WPT <- server()
  locs <- getStatus(WPT, "someID")
}
}

